RFC: A few functions to ease stdmod boilerplate
A common trend in criticizing the stdmod seems to be the amount of
repetitive boilerplate code that goes into init.pp.

This commit has two functions which reduce the boilerplate code in the
standard skeleton significantly. To be fair, the conf.pp could be using
the already existing pick from puppetlabs-stdlib as well.

The two functions should probably go to stdlib, and maybe deserve a better
name.